Prueba Habi/requirements.txt
mysql-connector-python>=8.0.32
pytest==7.4.0
fastapi>=0.68.0
sqlalchemy>=1.4.0
uvicorn>=0.15.0